Feat: 영농일지 아이디 Long 타입으로 변경

+++ core/data/src/main/java/kr/co/data/repository/DiaryRepositoryImpl.kt
@@ -43,7 +43,7 @@ internal class DiaryRepositoryImpl @Inject constructor(
             )
         }
 
-    override suspend fun getDiaryDetail(id: Int): DiaryEntity =
+    override suspend fun getDiaryDetail(id: Long): DiaryEntity =
         remote.fetchDetail(id).let{
             DiaryMapper.convert(it)
         }
@@ -79,7 +79,7 @@ internal class DiaryRepositoryImpl @Inject constructor(
     }
 
     override suspend fun updateDiary(
-        id: Int,
+        id: Long,
         crop:String,
         date: LocalDate,
         holidayList: List<HolidayEntity>,
@@ -108,7 +108,7 @@ internal class DiaryRepositoryImpl @Inject constructor(
         )
     }
 
-    override suspend fun deleteDiary(id: Int) {
+    override suspend fun deleteDiary(id: Long) {
         remote.delete(id)
     }
 }
