+++ core/remote/src/main/java/kr/co/remote/DiaryRemoteDataSourceImpl.kt
@@ -40,7 +40,7 @@ internal class DiaryRemoteDataSourceImpl @Inject constructor(
             }
             .let { flowOf(it) }
 
-    override suspend fun fetchDetail(id: Int): DiaryData =
+    override suspend fun fetchDetail(id: Long): DiaryData =
         client.get("$DIARY_URL/$id")
             .body<Dto<DiaryListResponse.DiaryResponse>>()
             .data.let { DiaryRemoteMapper.convert(it) }
@@ -107,7 +107,7 @@ internal class DiaryRemoteDataSourceImpl @Inject constructor(
     }
 
     override suspend fun update(
-        id: Int,
+        id: Long,
         crop: String,
         date: String,
         holidayList: List<HolidayData>,
@@ -148,7 +148,7 @@ internal class DiaryRemoteDataSourceImpl @Inject constructor(
         }
     }
 
-    override suspend fun delete(id: Int) {
+    override suspend fun delete(id: Long) {
         client.delete("$DIARY_URL/$id")
     }
